WebController: I added the mapping to test the api getting the data from the file
AreaMangager is used by the App.java
BricAreaManager is also used by App.java
this two are used so the api has a way to map to the database.

Area.java - is the way the databased is set up and this way Sprig puts the data into JSON automatically.

AreaRepository - This is where the queries for the database are created
Application.properties - this is where we connect the database to the Spring application